docs: clarify size of vm (#20)
If you create a template vm that is above the size that is specified for the nodes, proxmox will error because the provider will try to resize down to the vm size specified which is not supported. Currently the support node size defaults to 10G. Your template should be either less than or equal to that size.